hub/requirements.txt
# Hub requirements
fastapi>=0.109.0
uvicorn>=0.27.0
redis>=5.0.0
pyyaml>=6.0
python-multipart>=0.0.6
httpx>=0.26.0
